Caught Up
The wireframes are now complete to the same point they were at
yesterday, but without colors and images. "Next" buttons have been added
to each page, as an onSelectionChange wouldn't work unless there was a
null value selected at the outset, which isn't as clear to the user (or
at least I don't think it is - that's based on nothing but my own
supposition).