Registration Benefits
Added a modal window explaining benefits of registration. Also added a
similar modal to display Terms of Service, but it seems like it may not
be possible to add a text link to the label of a checkbox (which sort of
makes sense, as clicking a label is supposed to activate/deactivate the
checkbox).